Rename the test cases
