+++ src/test/kotlin/com/github/shiraji/model/FindPullRequestModelTest.kt
@@ -16,258 +16,246 @@ import git4idea.GitCommit
 import git4idea.history.GitHistoryUtils
 import git4idea.repo.GitRemote
 import git4idea.repo.GitRepository
+import io.mockk.*
+import io.mockk.impl.annotations.MockK
 import junit.framework.TestCase.*
 import org.junit.Before
-import org.junit.Ignore
 import org.junit.Test
-import org.junit.runner.RunWith
-import org.mockito.Matchers
-import org.mockito.Mock
-import org.mockito.Mockito.*
-import org.powermock.api.mockito.PowerMockito
-import org.powermock.core.classloader.annotations.PrepareForTest
-import org.powermock.modules.junit4.PowerMockRunner
-import java.util.*
 
-@RunWith(PowerMockRunner::class)
-@PrepareForTest(
-        GitHistoryUtils::class,
-        GitCommit::class,
-        ChangeListManager::class
-)
-@Ignore
 class FindPullRequestModelTest {
+
+    companion object {
+        private const val SELECTED_LINE = 100
+        private const val SELECTED_START_OFFSET = 100
+        private const val SELECTED_LINE_DIFF = 111
+        private const val SELECTED_END_OFFSET = 111
+        private const val HASH = "123"
+        private const val HASH_DIFF = "abc"
+        private const val PR_NUMBER = 10
+    }
 
     lateinit var model: FindPullRequestModel
 
-    @Suppress("MemberVisibilityCanBePrivate")
-    @Mock
-    lateinit var project: Project
-
-    @Suppress("MemberVisibilityCanBePrivate")
-    @Mock
-    lateinit var virtualFile: VirtualFile
-
-    @Suppress("MemberVisibilityCanBePrivate")
-    @Mock
-    lateinit var editor: Editor
-
-    @Suppress("MemberVisibilityCanBePrivate")
-    @Mock
-    lateinit var repository: GitRepository
-
-    @Suppress("MemberVisibilityCanBePrivate")
-    @Mock
-    lateinit var vcsRevisionNumber: VcsRevisionNumber
-
-    @Suppress("MemberVisibilityCanBePrivate")
-    @Mock
-    lateinit var virtualRoot: VirtualFile
-
-    @Suppress("MemberVisibilityCanBePrivate")
-    @Mock
+    @MockK
     lateinit var gitRepository: GitRepository
 
-    @Suppress("MemberVisibilityCanBePrivate")
-    @Mock
+    @MockK
     lateinit var changeListManager: ChangeListManager
 
-    @Suppress("MemberVisibilityCanBePrivate")
-    @Mock
-    lateinit var change: Change
-
-    @Suppress("MemberVisibilityCanBePrivate")
-    @Mock
+    @MockK
     lateinit var conf: PropertiesComponent
 
-    private val prNumber = 10
-    private val hashCode = "123"
-    private val diffHashCode = "abc"
-    private val selectedLine = 10
-    private val diffSelectedLine = 100
-    private val selectedOffline = 111
+    @MockK
+    lateinit var project: Project
 
-    private fun mockFindPullRequestCommit(results: List<GitCommit?>) {
-        PowerMockito.`when`(GitHistoryUtils.history(
-                Matchers.eq(project),
-                Matchers.eq(virtualRoot),
-                anyString(),
-                anyString(),
-                anyString(),
-                anyString(),
-                anyString()
-        )).thenReturn(results)
+    @MockK
+    lateinit var editor: Editor
+
+    @MockK
+    lateinit var virtualFile: VirtualFile
+
+    @MockK
+    lateinit var change: Change
+
+    @MockK
+    lateinit var selectionModel: SelectionModel
+
+    @MockK
+    lateinit var document: Document
+
+    @MockK
+    lateinit var vcsRevisionNumber: VcsRevisionNumber
+
+    @MockK
+    lateinit var virtualRoot: VirtualFile
+
+    @Before
+    fun setUp() {
+        MockKAnnotations.init(this)
+
+        model = FindPullRequestModel(project, editor, virtualFile, conf)
     }
 
-    private fun mockHistory(results: List<GitCommit>) {
-        PowerMockito.`when`(GitHistoryUtils.history(
-                Matchers.eq(project),
-                Matchers.eq(virtualRoot),
-                Matchers.anyString())
-        ).thenReturn(results)
-    }
-
-    private fun mockGetGitRepository(result: GitRepository = gitRepository) {
-        val remote = generateGitRemote(
-                name = "origin",
-                urls = listOf("git@github.com:shiraji/find-pull-request.git")
-        )
-        `when`(result.remotes).thenReturn(listOf(remote))
-    }
-
-    private fun mockIsUnversioned(result: Boolean) {
-        `when`(changeListManager.isUnversioned(virtualFile)).thenReturn(result)
-    }
-
-    private fun mockChangeType(result: Change.Type, ch: Change? = change) {
-        `when`(changeListManager.getChange(virtualFile)).thenReturn(ch)
-        ch?.let { `when`(it.type).thenReturn(result) }
-    }
-
-    private fun mockLineNumber(startLine: Int, endLine: Int) {
-        val selectionModel = mock(SelectionModel::class.java)
-        `when`(editor.selectionModel).thenReturn(selectionModel)
-        `when`(selectionModel.selectionStart).thenReturn(selectedOffline)
-        `when`(selectionModel.selectionEnd).thenReturn(selectedOffline)
-
-        val document = mock(Document::class.java)
-        `when`(editor.document).thenReturn(document)
-        `when`(document.getLineNumber(anyInt()))
-                .thenReturn(startLine)
-                .thenReturn(endLine)
-    }
-
-    private fun generateMockGitCommit(hashCode: String = "", fullMessage: String = ""): GitCommit {
-        val hash = generateMockHash(hashCode)
-        return PowerMockito.mock(GitCommit::class.java).also {
-            `when`(it.fullMessage).thenReturn(fullMessage)
-            `when`(it.commitTime).thenReturn(Date().time)
-            `when`(it.id).thenReturn(hash)
-        }
-    }
-
-    private fun generateMockHash(hashCode: String): Hash {
-        return mock(Hash::class.java).also {
-            `when`(it.asString()).thenReturn(hashCode)
-        }
+    private fun mockGitRepository(remotes: List<GitRemote>, root: VirtualFile = virtualRoot) {
+        every { gitRepository.remotes } returns remotes
+        every { gitRepository.root } returns root
     }
 
     private fun generateGitRemote(
             name: String = "origin",
-            urls: List<String> = listOf(""),
+            urls: List<String> = listOf("git@github.com:shiraji/find-pull-request.git"),
             pushUrls: Collection<String> = listOf(""),
             fetchRefSpecs: List<String> = listOf(""),
             pushRefSpecs: List<String> = listOf("")
     ) = GitRemote(name, urls, pushUrls, fetchRefSpecs, pushRefSpecs)
 
-    @Before
-    fun setup() {
-        model = FindPullRequestModel(project, editor, virtualFile, conf)
 
-        setUpForCreatePullRequestPath()
-        setUpForIsEnable()
+    private fun mockIsUnversioned(result: Boolean = false) {
+        every { changeListManager.isUnversioned(virtualFile) } returns result
+    }
+
+    private fun mockChangeType(result: Change.Type, ch: Change? = change) {
+        every { changeListManager.getChange(virtualFile) } returns ch
+        ch?.let { every { it.type } returns result }
+    }
+
+    private fun mockLineNumber(startLine: Int, endLine: Int) {
+        every { editor.selectionModel } returns selectionModel
+        every { selectionModel.selectionStart } returns SELECTED_START_OFFSET
+        every { selectionModel.selectionEnd } returns SELECTED_END_OFFSET
+        every { editor.document } returns document
+        every { document.getLineNumber(SELECTED_START_OFFSET) } returns startLine
+        every { document.getLineNumber(SELECTED_END_OFFSET) } returns endLine
+    }
+
+    private fun generateCommitId(hashCode: String): Hash {
+        return mockk {
+            every { this@mockk.asString() } returns hashCode
+        }
+    }
+
+    private fun generateGitCommit(hashCode: String = "", fullMessage: String = ""): GitCommit {
+        val hash = generateCommitId(hashCode)
+        return mockk {
+            every { this@mockk.fullMessage } returns fullMessage
+            every { this@mockk.commitTime } returns java.util.Date().time
+            every { this@mockk.id } returns hash
+        }
     }
 
     private fun mockConfig(isDisable: Boolean = false, isDebugMode: Boolean = false, isJumpToFile: Boolean = true, protocol: String = "https://") {
-        doReturn(isDisable).`when`(conf).isDisable()
-        doReturn(isDebugMode).`when`(conf).isDebugMode()
-        doReturn(isJumpToFile).`when`(conf).isJumpToFile()
-        doReturn(protocol).`when`(conf).getProtocol()
+        every { conf.isDisable() } returns isDisable
+        every { conf.isDebugMode() } returns isDebugMode
+        every { conf.isJumpToFile() } returns isJumpToFile
+        every { conf.getProtocol() } returns protocol
     }
 
-    private fun setUpForCreatePullRequestPath() {
-        PowerMockito.mockStatic(GitHistoryUtils::class.java)
-
-        `when`(repository.root).thenReturn(virtualRoot)
-        `when`(vcsRevisionNumber.asString()).thenReturn(hashCode)
+    private fun mockProject(isDisposed: Boolean = false) {
+        every { project.isDisposed } returns isDisposed
     }
 
-    private fun setUpForIsEnable() {
-        PowerMockito.mockStatic(ChangeListManager::class.java)
+    private fun mockRevisionNumber(hashCode: String) {
+        every { vcsRevisionNumber.asString() } returns hashCode
+    }
 
-        `when`(project.isDisposed).thenReturn(false)
+    private fun mockHistory(closestPRCommits: List<GitCommit>, mergeCommits: List<GitCommit>) {
+        mockkStatic(GitHistoryUtils::class)
 
-        mockConfig()
+        every {
+            GitHistoryUtils.history(
+                    project,
+                    virtualRoot,
+                    any(),
+                    any(),
+                    any(),
+                    any(),
+                    any())
+        } returns closestPRCommits
+
+        every {
+            GitHistoryUtils.history(project, virtualRoot, any())
+        } returns mergeCommits
     }
 
     @Test
     fun `Finding pull request`() {
-        val prCommit = generateMockGitCommit(fullMessage = "Merge pull request #$prNumber from")
-        mockFindPullRequestCommit(listOf(prCommit))
+        val prCommit1 = generateGitCommit(hashCode = HASH, fullMessage = "Merge pull request #$PR_NUMBER from")
+        val prCommit2 = generateGitCommit(hashCode = HASH_DIFF, fullMessage = "Merge pull request #${PR_NUMBER + 1} from")
+        val prCommit3 = generateGitCommit(hashCode = HASH_DIFF, fullMessage = "Merge pull request #${PR_NUMBER + 2} from")
 
-        val mergeCommit = generateMockGitCommit(hashCode = hashCode)
-        mockHistory(listOf(mergeCommit))
+        mockConfig()
+        mockGitRepository(listOf(generateGitRemote()))
+        mockHistory(closestPRCommits = listOf(prCommit1), mergeCommits = listOf(prCommit1, prCommit2, prCommit3))
+        mockRevisionNumber(HASH)
 
-        val path = model.createPullRequestPath(repository, vcsRevisionNumber)
-        assertEquals("pull/$prNumber/files", path)
+        val path = model.createPullRequestPath(gitRepository, vcsRevisionNumber)
+        assertEquals("pull/$PR_NUMBER/files", path)
     }
 
     @Test
     fun `Finding squash commit`() {
-        val prCommit = generateMockGitCommit()
-        mockFindPullRequestCommit(listOf(prCommit))
+        val prCommit1 = generateGitCommit(hashCode = HASH, fullMessage = "Foo (#$PR_NUMBER)")
 
-        val listOfCommits = generateMockGitCommit(fullMessage = "Foo (#$prNumber)")
-        // for findCommitLog
-        mockHistory(listOf(listOfCommits))
+        mockConfig()
+        mockGitRepository(listOf(generateGitRemote()))
+        mockHistory(closestPRCommits = emptyList(), mergeCommits = listOf(prCommit1))
 
-        val path = model.createPullRequestPath(repository, vcsRevisionNumber)
-        assertEquals("pull/$prNumber/files", path)
+        val path = model.createPullRequestPath(gitRepository, vcsRevisionNumber)
+        assertEquals("pull/$PR_NUMBER/files", path)
     }
 
     @Test(expected = NoPullRequestFoundException::class)
     fun `No PR found if no PR and the commit is not squash commit`() {
-        val prCommit = generateMockGitCommit()
-        mockFindPullRequestCommit(listOf(prCommit))
+        val prCommit1 = generateGitCommit(hashCode = HASH, fullMessage = "Merge pull request #$PR_NUMBER from")
 
-        val listOfCommits = generateMockGitCommit()
-        // for findCommitLog
-        mockHistory(listOf(listOfCommits))
+        mockConfig()
+        mockGitRepository(listOf(generateGitRemote()))
+        mockHistory(closestPRCommits = emptyList(), mergeCommits = listOf(prCommit1))
 
-        model.createPullRequestPath(repository, vcsRevisionNumber)
+        model.createPullRequestPath(gitRepository, vcsRevisionNumber)
         fail()
     }
 
     @Test
     fun `Found PR has different hash code but the commit is squash commit`() {
-        val prCommit = generateMockGitCommit(fullMessage = "Merge pull request #$prNumber from")
-        mockFindPullRequestCommit(listOf(prCommit))
+        val prCommit1 = generateGitCommit(hashCode = HASH_DIFF, fullMessage = "Merge pull request #${PR_NUMBER + 1} from")
+        val prCommit2 = generateGitCommit(hashCode = HASH_DIFF, fullMessage = "Foo (#${PR_NUMBER})")
+        val prCommit3 = generateGitCommit(hashCode = HASH, fullMessage = "Foo (#${PR_NUMBER})")
 
-        val mergeCommit = generateMockGitCommit(hashCode = diffHashCode)
-        mockHistory(listOf(mergeCommit))
+        mockConfig()
+        mockGitRepository(listOf(generateGitRemote()))
+        mockkStatic(GitHistoryUtils::class)
+        mockRevisionNumber(HASH)
 
-        val listOfCommits = generateMockGitCommit(fullMessage = "Foo (#$prNumber)")
-        // for findCommitLog
-        mockHistory(listOf(listOfCommits))
+        every {
+            GitHistoryUtils.history(
+                    project,
+                    virtualRoot,
+                    any(),
+                    any(),
+                    any(),
+                    any(),
+                    any())
+        } returns listOf(prCommit1)
 
-        val path = model.createPullRequestPath(repository, vcsRevisionNumber)
-        assertEquals("pull/$prNumber/files", path)
+        every {
+            GitHistoryUtils.history(project, virtualRoot, any())
+        } returnsMany listOf(listOf(prCommit2), listOf(prCommit3))
+
+        val path = model.createPullRequestPath(gitRepository, vcsRevisionNumber)
+        assertEquals("pull/$PR_NUMBER/files", path)
+        verify(exactly = 1) { GitHistoryUtils.history(project, virtualRoot, any(), any(), any(), any(), any()) }
+        verify(exactly = 2) { GitHistoryUtils.history(project, virtualRoot, any()) }
     }
 
     @Test(expected = NoPullRequestFoundException::class)
     fun `Found PR has different hash code and its commit message is not squash commit`() {
-        val prCommit = generateMockGitCommit(fullMessage = "Merge pull request #$prNumber from")
-        mockFindPullRequestCommit(listOf(prCommit))
+        val prCommit1 = generateGitCommit(hashCode = HASH_DIFF, fullMessage = "Merge pull request #${PR_NUMBER} from")
+        val prCommit2 = generateGitCommit(hashCode = HASH_DIFF, fullMessage = "Foo (#${PR_NUMBER})")
+        val prCommit3 = generateGitCommit(hashCode = HASH, fullMessage = "Merge pull request #${PR_NUMBER} from")
 
-        val mergeCommit = generateMockGitCommit(hashCode = diffHashCode)
-        mockHistory(listOf(mergeCommit))
+        mockConfig()
+        mockGitRepository(listOf(generateGitRemote()))
+        mockkStatic(GitHistoryUtils::class)
+        mockRevisionNumber(HASH)
 
-        val listOfCommits = generateMockGitCommit()
-        // for findCommitLog
-        mockHistory(listOf(listOfCommits))
+        every {
+            GitHistoryUtils.history(
+                    project,
+                    virtualRoot,
+                    any(),
+                    any(),
+                    any(),
+                    any(),
+                    any())
+        } returns listOf(prCommit1)
 
-        model.createPullRequestPath(repository, vcsRevisionNumber)
+        every {
+            GitHistoryUtils.history(project, virtualRoot, any())
+        } returnsMany listOf(listOf(prCommit2), listOf(prCommit3))
+
+        model.createPullRequestPath(gitRepository, vcsRevisionNumber)
         fail()
-    }
-
-    @Test
-    fun `isEnable true`() {
-        mockGetGitRepository()
-        mockIsUnversioned(false)
-        mockChangeType(Change.Type.MODIFICATION)
-        mockLineNumber(startLine = selectedLine, endLine = selectedLine)
-
-        assertTrue(model.isEnable(gitRepository, changeListManager))
     }
 
     @Test
@@ -275,57 +263,98 @@ class FindPullRequestModelTest {
         mockConfig(isDisable = true)
 
         assertFalse(model.isEnable(gitRepository, changeListManager))
+
+        verify(exactly = 1) { conf.isDisable() }
     }
 
     @Test
-    fun `isEnable true even the change is null`() {
-        mockGetGitRepository()
-        mockIsUnversioned(false)
-        mockChangeType(Change.Type.MODIFICATION, null)
-        mockLineNumber(startLine = selectedLine, endLine = selectedLine)
+    fun `isEnable false if project is disposed`() {
+        mockConfig()
+        mockProject(isDisposed = true)
+
+        assertFalse(model.isEnable(gitRepository, changeListManager))
+
+        verify(exactly = 1) { project.isDisposed }
+    }
+
+    @Test
+    fun `isEnable false if no git repository`() {
+        mockConfig()
+        mockProject()
+        mockGitRepository(emptyList())
+
+        assertFalse(model.isEnable(gitRepository, changeListManager))
+
+        verify { gitRepository.remotes }
+    }
+
+    @Test
+    fun `isEnable false if the file is not versioned`() {
+        mockConfig()
+        mockProject()
+        mockGitRepository(listOf(generateGitRemote()))
+        mockIsUnversioned(true)
+
+        assertFalse(model.isEnable(gitRepository, changeListManager))
+
+        verify(exactly = 1) { changeListManager.isUnversioned(virtualFile) }
+    }
+
+    @Test
+    fun `isEnable false if the change type is New`() {
+        mockConfig()
+        mockProject()
+        mockGitRepository(listOf(generateGitRemote()))
+        mockIsUnversioned()
+        mockChangeType(Change.Type.NEW)
+
+        assertFalse(model.isEnable(gitRepository, changeListManager))
+
+        verify(exactly = 1) { changeListManager.getChange(virtualFile) }
+    }
+
+    @Test
+    fun `isEnable false if multiple line is selected`() {
+        mockConfig()
+        mockProject()
+        mockGitRepository(listOf(generateGitRemote()))
+        mockIsUnversioned()
+        mockChangeType(Change.Type.MODIFICATION)
+        mockLineNumber(startLine = SELECTED_LINE, endLine = SELECTED_LINE_DIFF)
+
+        assertFalse(model.isEnable(gitRepository, changeListManager))
+
+        verify { document.getLineNumber(any()) }
+    }
+
+    @Test
+    fun `isEnable true`() {
+        mockConfig()
+        mockProject()
+        mockGitRepository(listOf(generateGitRemote()))
+        mockIsUnversioned()
+        mockChangeType(Change.Type.MODIFICATION)
+        mockLineNumber(startLine = SELECTED_LINE, endLine = SELECTED_LINE)
 
         assertTrue(model.isEnable(gitRepository, changeListManager))
     }
 
     @Test
-    fun `isEnable false if no git repository`() {
-        assertFalse(model.isEnable(gitRepository, changeListManager))
-    }
+    fun `isEnable true even the change is null`() {
+        mockConfig()
+        mockProject()
+        mockGitRepository(listOf(generateGitRemote()))
+        mockIsUnversioned()
+        mockChangeType(Change.Type.MODIFICATION, null)
+        mockLineNumber(startLine = SELECTED_LINE, endLine = SELECTED_LINE)
 
-    @Test
-    fun `isEnable false if the file is not versioned`() {
-        mockGetGitRepository()
-        mockIsUnversioned(true)
-
-        assertFalse(model.isEnable(gitRepository, changeListManager))
-    }
-
-    @Test
-    fun `isEnable false if the change type is New`() {
-        mockGetGitRepository()
-        mockIsUnversioned(false)
-        mockChangeType(Change.Type.NEW)
-
-        assertFalse(model.isEnable(gitRepository, changeListManager))
-    }
-
-    @Test
-    fun `isEnable false if multiple line is selected`() {
-        mockGetGitRepository()
-        mockIsUnversioned(false)
-        mockChangeType(Change.Type.MODIFICATION)
-        mockLineNumber(startLine = selectedLine, endLine = diffSelectedLine)
-
-        assertFalse(model.isEnable(gitRepository, changeListManager))
+        assertTrue(model.isEnable(gitRepository, changeListManager))
     }
 
     @Test
     fun `createWebRepoUrl returns correct https url if there is upstream url`() {
-        val remote = generateGitRemote(
-                name = "upstream",
-                urls = listOf("git@github.com:shiraji/find-pull-request.git")
-        )
-        `when`(gitRepository.remotes).thenReturn(listOf(remote))
+        mockConfig()
+        mockGitRepository(listOf(generateGitRemote(name = "upstream")))
 
         val result = model.createWebRepoUrl(gitRepository)
 
@@ -334,11 +363,8 @@ class FindPullRequestModelTest {
 
     @Test
     fun `createWebRepoUrl returns correct https url if there is origin url`() {
-        val remote = generateGitRemote(
-                name = "origin",
-                urls = listOf("git@github.com:shiraji/find-pull-request.git")
-        )
-        `when`(gitRepository.remotes).thenReturn(listOf(remote))
+        mockConfig()
+        mockGitRepository(listOf(generateGitRemote()))
 
         val result = model.createWebRepoUrl(gitRepository)
 
@@ -347,11 +373,10 @@ class FindPullRequestModelTest {
 
     @Test
     fun `createWebRepoUrl returns correct https url if url is https format`() {
-        val remote = generateGitRemote(
-                name = "origin",
+        mockConfig()
+        mockGitRepository(listOf(generateGitRemote(
                 urls = listOf("https://github.com/shiraji/find-pull-request.git")
-        )
-        `when`(gitRepository.remotes).thenReturn(listOf(remote))
+        )))
 
         val result = model.createWebRepoUrl(gitRepository)
 
@@ -360,11 +385,10 @@ class FindPullRequestModelTest {
 
     @Test
     fun `createWebRepoUrl returns correct https url if url is custom domain`() {
-        val remote = generateGitRemote(
-                name = "origin",
+        mockConfig()
+        mockGitRepository(listOf(generateGitRemote(
                 urls = listOf("https://github.enterprise.local/shiraji/find-pull-request.git")
-        )
-        `when`(gitRepository.remotes).thenReturn(listOf(remote))
+        )))
 
         val result = model.createWebRepoUrl(gitRepository)
 
@@ -373,11 +397,11 @@ class FindPullRequestModelTest {
 
     @Test
     fun `createWebRepoUrl returns null if there is no remote for origin or upstream`() {
-        val remote = generateGitRemote(
+        mockConfig()
+        mockGitRepository(listOf(generateGitRemote(
                 name = "shiraji",
                 urls = listOf("https://github.enterprise.local/shiraji/find-pull-request.git")
-        )
-        `when`(gitRepository.remotes).thenReturn(listOf(remote))
+        )))
 
         val result = model.createWebRepoUrl(gitRepository)
 
@@ -386,11 +410,10 @@ class FindPullRequestModelTest {
 
     @Test
     fun `createWebRepoUrl returns null if the format is no user name`() {
-        val remote = generateGitRemote(
-                name = "origin",
+        mockConfig()
+        mockGitRepository(listOf(generateGitRemote(
                 urls = listOf("https://github.enterprise.local/shiraji/find-pull-request")
-        )
-        `when`(gitRepository.remotes).thenReturn(listOf(remote))
+        )))
 
         val result = model.createWebRepoUrl(gitRepository)
 
@@ -399,11 +422,10 @@ class FindPullRequestModelTest {
 
     @Test
     fun `createWebRepoUrl returns correct https url if url is https format for bitbucket`() {
-        val remote = generateGitRemote(
-                name = "origin",
+        mockConfig()
+        mockGitRepository(listOf(generateGitRemote(
                 urls = listOf("https://shiraji@bitbucket.org/shiraji/find-pull-request.git")
-        )
-        `when`(gitRepository.remotes).thenReturn(listOf(remote))
+        )))
 
         val result = model.createWebRepoUrl(gitRepository)
 
@@ -412,11 +434,10 @@ class FindPullRequestModelTest {
 
     @Test
     fun `createWebRepoUrl returns correct https url if url is https format for gitlab`() {
-        val remote = generateGitRemote(
-                name = "origin",
+        mockConfig()
+        mockGitRepository(listOf(generateGitRemote(
                 urls = listOf("https://gitlab.com/shiraji/find-pull-request.git")
-        )
-        `when`(gitRepository.remotes).thenReturn(listOf(remote))
+        )))
 
         val result = model.createWebRepoUrl(gitRepository)
